requirements.txt
google-api-python-client
google-auth
oauth2client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
beautifulsoup4
sentence-transformers
